🏗️ (new): Create view model for displaying food list view
Create a view model for the TreatsListView, where the functionality for making the actual network call and getting & storing the decoded JSON list of food is defined. The view model is then initialized in its associated view file to listen for any changes and re-renders the view with the populated food data on-screen.