🏗️ (new): Update view model with error cases and display it to users
Added the error cases to the NetworkManager failure result case, in the event if network calls to retrieve and decode the JSON data fails. Depending on the failure case that trigger, the error alert details will be published to the associated view to trigger the display of the alert window to the user by informing them what went wrong.