🚸 (new): Display placeholder food image if image data is missing
Created an image loader feature where the display images of the food will show a placeholder image in the event before the network call to retrieve the actual images data is completed or have failed. If image data retrieval is successful from the server, the actual food image will be displayed.