💄 (new): Add UI layout of food detail view pop-up
Created a new file - TreatDetailView that shows the details of the selected food. This detail screen will be displayed to users in the form of a pop-up view and a button at the top-right to dismiss it. A button for users to add the selected food to their order screen is also included in this detail view.