✨ (new): Set food detail view to display on user tap
Added the functionality where users tapped on any of the food in the list, a detail view of that selected food will display on screen like a popup view, with a blur effect set on the background. For UX optimization, interaction on the background is disabled until the display of the food detail view is dismissed by the user.