✨ (new): Add functionality to dimiss button and change image display
Implemented the functionality to the dismiss button by setting a @Binding wrapper to the @State property in the TreatListView file to listen for changes on whether to display or dimiss the detail view. Also, replaced the previous placeholder food image to display the correct one based on the selected food item.