🏗️ (new): Add an Order object to initialize as @EnvironmentObject
Created a Order object that holds a list of food items that gets added to the order view. This object will serve as an @EnvironmentObject that will be initialized in the @main TreatsApp file, because the Order object needs to be accessed by two child views - TreatDetailView & OrderView for some their core features to work, such as adding items to order from the food details pop-up window & deleting items from the order screen.