💄 (new): Add items count badge to display on Order bag icon
Added the .badge() modifier to the OrderView tab icon to display the number of item(s) in the Order screen. Also, refactored the tab items to use a Label instead of the previous Image & Text view. Fixed an Xcode preview crashing bug in the TreatsTabView_Previews object by initializing the Order environment object.