🚸 (chore): Add focus state to text fields and keyboard dismiss button
Added form field focus functionality using @FocusState wrapper to determine the selected text field in the user account form. The flow of the user input on the form has been improved by auto-directing the user to the next input field when the user tapped on the "next" button. After all three input fields have been filled, the user can simply tap the "continue" button to dismiss the keyboard. A dimiss toolbar button was also added to the keyboard as an alternative for users to dimiss it anytime.