🚸 (new): Set DatePicker range for user age of 18-110 y/o
Added an extension to the Date object that dynamically calculates the exact dates for 18 years ago and 110 years ago from today's date. These date calculations is then used to set a limited range on the user birthdate DatePicker in the account's screen. This change improves the UX by limiting the user's need to scroll through many years on the DatePicker to get their birhtdate and also prevent the unrealistic input of dates such as a birthdate from the year 1200.